fix(kucoin): parse futures position leverage as a decimal
KuCoin reports `leverage` on `/api/v1/positions` as a decimal — cross
positions return the effective leverage (e.g. 9.51) — but the payload was
bound to an int64 field, so json.Unmarshal rejected every response with
"cannot unmarshal number 9.51 into Go struct field
futures_Position.data.leverage of type int64" and the gateway answered 400.

Bind the field to float64 and stringify it with FloatToStringAll, matching
the other decimal fields on the same struct.

+++ kucoin/futures_getPositions.go
@@ -50,7 +50,7 @@ type futures_Position struct {
 	RealisedPnl      float64 `json:"realisedPnl"`
 	UnrealisedPnl    float64 `json:"unrealisedPnl"`
 	AvgEntryPrice    float64 `json:"avgEntryPrice"`
-	Leverage         int64   `json:"leverage"`
+	Leverage         float64 `json:"leverage"`
 	OpeningTimestamp int64   `json:"openingTimestamp"`
 	CurrentTimestamp int64   `json:"currentTimestamp"`
 }
